+++ src/components/currency/Home.js
@@ -1,25 +1,23 @@
+import React, { useCallback } from 'react';
 import { useDispatch, useSelector } from 'react-redux';
-import { useState } from 'react';
 import Exchange from './Exchange';
 import { filterdata } from '../../redux/exchangeSlice';
 
 const Home = () => {
   const dispatch = useDispatch();
-  const { exchange } = useSelector((state) => state.exchange);
-  const limitValues = Object.values(exchange).slice(0, 200);
-  const [result, setResult] = useState('');
+  const exchange = useSelector((state) => state.exchange);
 
-  const search = (e) => {
+  const search = useCallback((e) => {
     const { value } = e.target;
-    const filterData = limitValues.filter((item) => item.currencyCode === value.toUpperCase());
+    const filterData = exchange.slice(0, 200).filter((item) => item.currencyCode
+    === value.toUpperCase());
     dispatch(filterdata(filterData));
-    setResult(value);
-  };
+  }, [dispatch, exchange]);
 
   return (
     <section className="Container">
       <div className="HomeTop">
-        <input onChange={search} type="search" value={result} placeholder="Search" />
+        <input onChange={search} type="search" placeholder="Search" />
         <h1>Exchange Your Currency</h1>
       </div>
       <Exchange />
